orchestrator/requirements.txt
python-dotenv>=1.0
openpyxl>=3.1
openai>=1.0
google-genai>=1.0
flask>=3.0
mcp>=1.0
